// src online/types/general.ts
export type AppState = FeatureUserLoginState &
    T_FeatureCustomerReducerState &
    T_SmeLoan_FeatureAccountState &
    T_FeatureDocumentState &
    FeatureNewsState<FeatureUserLoginState> &
    FeatureNoticesState &
    FeatureTranslationStaticContentState<FeatureUserLoginState> &
    FeatureNoticesState &
    T_FeatureInvoicesState &
    T_SmeFeatureTransactionsState &
    FeatureInitializerState &
    FeatureSmeWithdrawState<FeatureUserLoginState> &
    FeatureLoaderState &
    T_FeatureLoginSessionState &
    T_FeaturePageInitializerState &
    T_FeatureReportingState &
    T_FeatureSessionState &
    T_KycReducerState;
// ...
export enum E_Routes {
    ROOT = "/",
    LOGIN = "/login",
    ACCOUNTS = "/chooseaccount",
    USER = "/user",
    FRONT = "/front",
    LOAN = "/loan",
    CONTACT = "/contact",
    CHANGE = "/change",
    TOPUP = "/topup",
    LOGOUT = "/logout",
    ERROR = "/error",
    NO_LOAN = "/noloan",
    EXPIRED = "/expired",
    CHOOSE_ACCOUNT = "/chooseaccount",
    NOT_FOUND = "/notfound",
    ALL_OTHERS = "*",
    KYC = "/application/:id",
    APPLICATION = "/application",
}